+++ Gauss/test/tst_gausstests.h
@@ -9,7 +9,7 @@ class GaussTests : public QObject
     Q_OBJECT
 private slots:
     void int2x2();
-    void float2x2();
+    void double2x2();
     void inconsistent2x2();
 
     void inconsistentRealCase();
