+++ common/include/common.h
@@ -4,13 +4,13 @@
 #include <functional>
 #include <map>
 
-using Coord = float;
-using CoordDiff = float;
-using MathFunc = std::function<float (Coord, Coord)>;
+using Coord = double;
+using CoordDiff = double;
+using MathFunc = std::function<double (Coord, Coord)>;
 
-using BoundaryCond1 = std::map<int, float>;
+using BoundaryCond1 = std::map<int, double>;
 
-using Vector = std::vector<float>;
+using Vector = std::vector<double>;
 using Matrix = std::vector<Vector>;
 
 Matrix create2DMatrix(size_t width, size_t height);
